server/requirements.txt
appdirs==1.4.4
asgiref==3.3.4
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cssselect==1.1.0
Django==3.1.7
django-cors-headers==3.7.0
django-notifications-hq==1.6.0
django-sendgrid-v5==0.9.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
drf-yasg==1.20.0
fake-useragent==0.1.11
feedparser==6.0.2
idna==2.10
importlib-metadata==2.1.1
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
lxml==4.6.3
MarkupSafe==1.1.1
numpy==1.20.2
packaging==20.9
pandas==1.2.4
parse==1.19.0
psycopg2-binary==2.8.6
pycodestyle==2.7.0
pyee==8.1.0
PyJWT==2.0.1
pyparsing==2.4.7
pyppeteer==0.2.5
pyquery==1.4.3
python-dateutil==2.8.1
python-dotenv==0.17.0
pytz==2021.1
requests==2.25.1
requests-html==0.10.0
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
sgmllib3k==1.0.0
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
toml==0.10.2
tqdm==4.60.0
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.4
w3lib==1.22.0
websockets==8.1
yahoo-fin==0.8.8
zipp==3.4.1
flake8
flake8-broken-line
flake8-commas
flake8-isort
